bluetooth: only set background in bluetooth skill after speak TTS
PR-URL: https://github.com/yodaos-project/yodart/pull/516

Fix bug:
https://bug.rokid-inc.com/zentaopms/www/index.php?m=bug&f=view&bugID=19054
https://bug.rokid-inc.com/zentaopms/www/index.php?m=bug&f=view&bugID=18875

Because for the two scenes of Bluetooth phone & Bluetooth music, there is a setting of keepAlive=true to prevent it from being killed by system. After this setting, it will not receive onDestroy but only onBackground. And in lifecycle onBackground we check if current skill is bluetooth phone or bluetooth music, then disconnect the connection (This is according product definition).

However, after speak TTS such as `已连接上你的xxx`, we will call setBackground manually to let bluetooth goto background like a service. Therefore, during a call process, if bluetooth is opened and the relevant TTS is spoken. Then, after the bluetooth connection is successful, the Bluetooth phone skill will be activated immediately. Thereafter, when the TTS broadcast is completed, it will actively setBackground, thus causing the lifecycle function onBackground to be received in the Bluetooth phone skill, and then the Bluetooth is actively disconnected.